requirements/app.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/app.in
#
altair==4.2.0
    # via streamlit
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
astor==0.8.1
    # via streamlit
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via
    #   jsonschema
    #   streamlit
backcall==0.2.0
    # via ipython
backports.zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
base58==2.1.1
    # via streamlit
black==21.12b0
    # via ipython
bleach==4.1.0
    # via nbconvert
blinker==1.4
    # via streamlit
cachetools==5.0.0
    # via streamlit
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via argon2-cffi-bindings
charset-normalizer==2.0.10
    # via requests
click==8.0.3
    # via
    #   black
    #   streamlit
debugpy==1.5.1
    # via ipykernel
decorator==5.1.1
    # via
    #   ipython
    #   validators
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.3
    # via
    #   altair
    #   jupyter-client
    #   nbconvert
executing==0.8.2
    # via stack-data
gitdb==4.0.9
    # via gitpython
gitpython==3.1.26
    # via streamlit
idna==3.3
    # via requests
importlib-resources==5.4.0
    # via jsonschema
ipykernel==6.7.0
    # via
    #   ipywidgets
    #   notebook
    #   pydeck
ipython==8.0.1
    # via
    #   ipykernel
    #   ipywidgets
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   nbformat
    #   notebook
ipywidgets==7.6.5
    # via pydeck
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via
    #   altair
    #   nbconvert
    #   notebook
    #   pydeck
jsonschema==4.4.0
    # via
    #   altair
    #   nbformat
jupyter-client==7.1.2
    # via
    #   ipykernel
    #   nbclient
    #   notebook
jupyter-core==4.9.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.2
    # via ipywidgets
markupsafe==2.0.1
    # via jinja2
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
mypy-extensions==0.4.3
    # via black
nbclient==0.5.10
    # via nbconvert
nbconvert==6.4.0
    # via notebook
nbformat==5.1.3
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.4
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.7
    # via widgetsnbextension
numpy==1.22.1
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==21.3
    # via
    #   bleach
    #   streamlit
pandas==1.4.0
    # via
    #   altair
    #   streamlit
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.0.0
    # via streamlit
platformdirs==2.4.1
    # via black
prometheus-client==0.12.0
    # via notebook
prompt-toolkit==3.0.24
    # via ipython
protobuf==3.19.3
    # via streamlit
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pyarrow==6.0.1
    # via streamlit
pycparser==2.21
    # via cffi
pydeck==0.7.1
    # via streamlit
pygments==2.11.2
    # via
    #   ipython
    #   jupyterlab-pygments
    #   nbconvert
pympler==1.0.1
    # via streamlit
pyparsing==3.0.7
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pysqlite3==0.4.6
    # via -r requirements/app.in
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   pandas
    #   streamlit
pytz==2021.3
    # via pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   notebook
requests==2.27.1
    # via streamlit
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   validators
smmap==5.0.0
    # via gitdb
stack-data==0.1.4
    # via ipython
streamlit==1.4.0
    # via -r requirements/app.in
terminado==0.12.1
    # via notebook
testpath==0.5.0
    # via nbconvert
toml==0.10.2
    # via streamlit
tomli==1.2.3
    # via black
toolz==0.11.2
    # via altair
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   streamlit
    #   terminado
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   pydeck
typing-extensions==4.0.1
    # via black
tzdata==2021.5
    # via pytz-deprecation-shim
tzlocal==4.1
    # via streamlit
urllib3==1.26.8
    # via requests
validators==0.18.2
    # via streamlit
watchdog==2.1.6
    # via streamlit
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
widgetsnbextension==3.5.2
    # via ipywidgets
zipp==3.7.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
